Add artillery config and random number generator for CSV. Move legacy spec files.
