+++ spec/legacy-spec/dummyData.js
@@ -1,0 +1,120 @@
+const dummyData = {
+  sample1:
+    [
+      {
+          "_id": "5f2f132dbe15a9c58f3dc0be",
+          "rating": 3,
+          "author": "Fredrick King",
+          "date": "2020-07-01T12:34:52.733Z",
+          "description": "Et cumque quis. Molestias rerum aut iusto. Porro cum aspernatur esse odio harum excepturi. Fugit dolorem consectetur autem. Quam alias est nemo veritatis aut voluptate facilis facere blanditiis. Est consectetur est veniam quisquam rem maxime odit.",
+          "listing_id": "5f2f132dbe15a9c58f3dc06f",
+          "listingName": "Refined Metal Guitar"
+      },
+      {
+          "_id": "5f2f132dbe15a9c58f3dc0bd",
+          "rating": 2,
+          "author": "Riley Osinski",
+          "date": "2020-04-01T19:45:20.498Z",
+          "description": "Molestiae ea ipsum exercitationem. Et quia accusamus. Nihil corporis soluta possimus debitis ea atque laborum. Qui delectus nemo sint.",
+          "listingName": "Refined Metal Guitar",
+          "listing_id": "5f2f132dbe15a9c58f3dc052"
+      },
+      {
+          "_id": "5f2f132dbe15a9c58f3dc0bc",
+          "rating": 3,
+          "author": "Hayden Jacobson",
+          "date": "2019-12-10T18:21:21.747Z",
+          "description": "Impedit aspernatur rerum enim non quia sapiente praesentium molestiae. Non dolor quidem sit ut eius ut esse. Qui perspiciatis non sint et non modi ab veritatis. Vel cum quis blanditiis voluptates unde animi est. Qui natus officia nemo sed quos.",
+          "listingName": "Refined Metal Guitar",
+          "listing_id": "5f2f132dbe15a9c58f3dc06f"
+      },
+      {
+          "_id": "5f2f132dbe15a9c58f3dc0bb",
+          "rating": 4,
+          "author": "Santiago Hamill",
+          "date": "2019-12-08T20:59:45.191Z",
+          "description": "Dolore a dignissimos sapiente cumque alias nihil amet. Aut voluptas sapiente iusto eos architecto accusamus. Hic ut reiciendis sed velit optio pariatur. Qui quia asperiores inventore adipisci fuga occaecati iste perspiciatis pariatur. Autem rerum ullam recusandae explicabo minus esse officia autem rerum.",
+          "listingName": "Refined Metal Guitar",
+          "listing_id": "5f2f132dbe15a9c58f3dc052"
+      }
+    ],
+  sample1AverageRating: 3,
+  sample2:
+    [
+      {
+          "_id": "5f2f132dbe15a9c58f3dc0b2",
+          "rating": 3,
+          "author": "Nicolette Sauer",
+          "date": "2020-04-25T19:23:58.909Z",
+          "description": "Culpa molestias atque. Doloremque non quae. Cupiditate magnam cumque a nihil omnis. Recusandae dicta fuga quaerat.",
+          "listingName": "Refined Metal Guitar",
+          "listing_id": "5f2f132dbe15a9c58f3dc046"
+      },
+      {
+          "_id": "5f2f132dbe15a9c58f3dc0ad",
+          "rating": 4,
+          "author": "Will Corwin",
+          "date": "2020-02-23T21:15:22.680Z",
+          "description": "Animi quas provident labore et sint nisi modi. Nesciunt qui modi. Et ut veniam. Non aliquid exercitationem itaque. Mollitia omnis qui et.",
+          "listingName": "Refined Metal Guitar",
+          "listing_id": "5f2f132dbe15a9c58f3dc0a1"
+      },
+      {
+          "_id": "5f2f132dbe15a9c58f3dc0af",
+          "rating": 4,
+          "author": "Nicole O'Kon",
+          "date": "2020-01-22T17:02:09.878Z",
+          "description": "Aut error est. Quis nihil laboriosam velit qui eveniet quas. Consequuntur commodi veritatis inventore rem a soluta reiciendis quae architecto. Explicabo voluptatem voluptatem asperiores veritatis id ut dolore quibusdam. Eos non eius.",
+          "listingName": "Refined Metal Guitar",
+          "listing_id": "5f2f132dbe15a9c58f3dc0a1"
+      },
+      {
+          "_id": "5f2f132dbe15a9c58f3dc0ae",
+          "rating": 1,
+          "author": "Tina Effertz",
+          "date": "2019-12-30T18:40:48.311Z",
+          "description": "Ut sit quod est omnis eaque eius. Esse dolores temporibus ullam. Illo alias quidem et ea harum. At pariatur incidunt molestiae quas id fugiat dolor. Illum tempora velit voluptatum ex at est nesciunt. Amet fuga esse praesentium corrupti.",
+          "listingName": "Refined Metal Guitar",
+          "listing_id": "5f2f132dbe15a9c58f3dc075"
+      },
+      {
+          "_id": "5f2f132dbe15a9c58f3dc0b0",
+          "rating": 2,
+          "author": "Jess Hirthe",
+          "date": "2019-12-12T00:16:37.663Z",
+          "description": "Consectetur eos vitae. Consequatur sint ex ex nobis consequatur possimus. Laudantium eos illo rem quia consequatur et aut voluptatem recusandae. Aut id saepe incidunt. Exercitationem suscipit consequatur non fugit dicta exercitationem.",
+          "listingName": "Refined Metal Guitar",
+          "listing_id": "5f2f132dbe15a9c58f3dc075"
+      },
+      {
+          "_id": "5f2f132dbe15a9c58f3dc0b4",
+          "rating": 1,
+          "author": "Mr. Bette Ankunding",
+          "date": "2019-11-29T18:42:46.210Z",
+          "description": "Sunt expedita tempore adipisci sunt illum. Voluptatum hic sit ducimus minima eum vero. Harum quia nemo sapiente blanditiis quidem. Dolor alias voluptas non omnis nesciunt architecto dolor odit et.",
+          "listingName": "Refined Metal Guitar",
+          "listing_id": "5f2f132dbe15a9c58f3dc0a1"
+      },
+      {
+          "_id": "5f2f132dbe15a9c58f3dc0b3",
+          "rating": 4,
+          "author": "Dr. Horacio Hane",
+          "date": "2019-10-29T05:56:35.323Z",
+          "description": "Ad consequuntur atque. Incidunt aut omnis architecto dolorem totam expedita nulla adipisci nemo. Ipsum qui dolorum et dolores quis numquam fuga rem. Sint et temporibus laboriosam error aut.",
+          "listingName": "Refined Metal Guitar",
+          "listing_id": "5f2f132dbe15a9c58f3dc0a1"
+      },
+      {
+          "_id": "5f2f132dbe15a9c58f3dc0b1",
+          "rating": 3,
+          "author": "Hester Rogahn",
+          "date": "2019-10-26T09:57:39.878Z",
+          "description": "Repellat quisquam vitae odit occaecati natus totam aut voluptas quidem. Ut voluptatem excepturi doloremque earum reprehenderit esse. Qui ab voluptatem aut. Excepturi dolores impedit.",
+          "listingName": "Refined Metal Guitar",
+          "listing_id": "5f2f132dbe15a9c58f3dc075"
+      }
+    ],
+  sample2AverageRating: 2.75
+}
+
+export default dummyData;
